Cast the output from strtoul to unsigned int
On platforms where unsigned long is 64 bit, the implicit conversion
to unsigned int can warn if such warnings are enabled.

Add casts to incidate that this is known/intended.

+++ tests/checkasm/checkasm.c
@@ -521,7 +521,7 @@ int main(int argc, char *argv[]) {
             fprintf(stderr, "]\n");
             return 0;
         } else {
-            state.seed = strtoul(argv[1], NULL, 10);
+            state.seed = (unsigned int) strtoul(argv[1], NULL, 10);
         }
 
         argc--;
